update selector, move machine selector to combined.js
