+++ src/selectors/machine/index.js
@@ -1,1 +1,1 @@
-export * from './selectors';
+export * from './combined';
