+++ src/utils/status/host/hostStatus.js
@@ -2,7 +2,7 @@ import {
   getOperationalStatus,
   getProvisioningState,
   getHostErrorMessage,
-  getNodeUnschedulable,
+  isNodeUnschedulable,
 } from '../../../selectors';
 
 import { HOST_STATUS_TO_TEXT, HOST_STATUS_READY, HOST_STATUS_REGISTERING } from './constants';
@@ -23,5 +23,5 @@ export const getHostStatus = host => {
 export const getSimpleHostStatus = host => getHostStatus(host).status;
 
 export const canHostAddMachine = host => [HOST_STATUS_READY].includes(getSimpleHostStatus(host));
-export const canHostStartMaintenance = hostNode => hostNode && !getNodeUnschedulable(hostNode);
-export const canHostStopMaintenance = hostNode => hostNode && getNodeUnschedulable(hostNode);
+export const canHostStartMaintenance = hostNode => hostNode && !isNodeUnschedulable(hostNode);
+export const canHostStopMaintenance = hostNode => isNodeUnschedulable(hostNode);
